+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Setting of `contentType` to use detection when content type is not provided
+
+### Changed
+
 - **[BREAKING]** updated `FormSubmissionApprovalsResponse` type
 
 ## [15.0.0] - 2024-09-03